Bug 2914, make SSL.Context() work with SWIG 1.3.24. Patch from Robin Dunn.
git-svn-id: http://svn.osafoundation.org/chandler/trunk@5197 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ internal/m2crypto/M2Crypto/SSL/Context.py
@@ -18,13 +18,13 @@ class _ctxmap:
         self.map = {}
 
     def __getitem__(self, key):
-        return self.map[key] 
+        return self.map[long(key)] 
 
     def __setitem__(self, key, value):
-        self.map[key] = value
+        self.map[long(key)] = value
 
     def __delitem__(self, key):
-        del self.map[key]
+        del self.map[long(key)]
 
 def map():
     if _ctxmap.singleton is None:
